# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Application Assistant",
    "version": "0.1.0"
  },
  "paths": {
...
    "/job-ads/analyse": {
      "post": {
        "tags": [
          "job-ads"
        ],
        "summary": "Analyse Job Ad",
        "description": "Reads an advert. Stores nothing - the result is prefill, and the user has not agreed to it.",
        "operationId": "analyse_job_ad_job_ads_analyse_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobAdText"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/job-ads/resume": {
      "post": {
        "tags": [
          "job-ads"
        ],
        "summary": "Write Draft Resume",
        "description": "Writes a CV for an advert with no application yet. Stores nothing: the create form carries\nthe draft in a hidden field and it is written with the application, or not at all.",
        "operationId": "write_draft_resume_job_ads_resume_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobAdText"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobAdText": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "JobAdText"
      },
      "JobAnalysis": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "match_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Rating"
          },
          "match_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Summary"
          },
          "match_strengths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Strengths"
          },
          "match_weaknesses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Weaknesses"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company",
          "sector",
          "location",
          "match_rating",
          "match_summary",
          "match_strengths",
          "match_weaknesses"
        ],
        "title": "JobAnalysis",
        "description": "What one model call returns. Also the response body of POST /job-ads/analyse.\n\nThe match fields are null when there is no profile to score against. No defaults: OpenAI's\nstrict structured outputs require every property to be required, and nullability is how\n\"absent\" is expressed. The 210-character summary and the four-entry lists are prompt rules,\nnot constraints here - a length limit in this model would turn a chatty model into a 500."
      },
...
      "ResumeDraft": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ResumeDraft",
        "description": "A tailored CV for an advert that has no application yet. Stored nowhere - the create form\ncarries it until the application is written."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}